(#6) reorganized forex app structure and renamed
