+++ forex/pkg/exchange/streamer.go
@@ -1,4 +1,4 @@
-package extract
+package exchange
 
 import "github.com/gorilla/websocket"
