+++ forex/pkg/exchange/trader_made.go
@@ -1,4 +1,4 @@
-package extract
+package exchange
 
 import (
 	"fmt"
